Wireframe Layout Adjustment
Removed two featurettes on the main page, and added another row of container marketing to the bottom. Changed the words in the navbar to Home, Shop, About, and Contact. Added a Account and Cart Icon to the navbar as placeholders. Added new navbar to the other files, removed two featurettes on the about page.