Navbar + Carousel Update
Added image to the Carousel and in the img file, changed the placeholder text, and rearranged the text position in the carousel. Added the script for bootstrap JS in the header. Changed the icon and text color in the navbar, and added two more icons, a filled heart and card. and lastly, adjusted the position for the images in the carousel.