About Page Start and Adjustments
Replaced the placeholder navbar example in most pages to the new one done in the home page. Added the images to the placeholder image box, and replaced the placeholder text to the information given from the javacoffee.com webpage. added a about.css for the about.html for adjustments and customization. added padding to the body and the featurette's text to make the layout more balanced. and replaced the other page's head to the index.html head, that includes all the links missing.