Icon Update
Added a modal for all the icon in the nav bar; account, cart, wishlist, and gift card. Added a form in the account and gift card modal. changhed the primary button color to a darker coffee brown. added a sign-up.html file for the account. and applied the chanhged made in the index.html to the rest of the .html documents.