Footer Changes
Changed the contents in the footer to include a heading, a short text, and a form to input email for newsletter. Also added a instagram, facebook, and twitter icon at the bottom. adjusted the width to fill the whole width of the screen. Changed the background color to the same one as the nav bar. changed the text to white, and applied the css and html of the footer to the rest of the .html files that is in the nav bar; about, contact, shop, and account.