Product Page Update
Added more images for the product page from unsplash, and removed/replaced other images used in the main page with the new images. changed the color for the product "boxes". changed the placeholder text to actual names and information. added a background image for the heading at the top. adjusted the box around the product and the image to balance the size. and changed the border radius from pointed to slightly rounded.