Small Adjustments
Adjusted the spacing of the images in the about page to align with the text more, and adjusted the width of one of the paragraph in the about page. And adjusted the image of the shop location's spacing from the border of the site so its more centered with the text.