Modal Update
Added modals for the products on the main page and the product page. Added the image, body text, and price in the modal body for all, and added a 'add to cart' and a 'add to wishlist' button in the modal footer. removed the close button from the modal footer. applied the toggle code to trigger modal for every button for the products.